LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-2-expected.txt

⨀

⨁

⨃

⨅

⨇

⨈

⨉

⨊

⫼

⫿

⨀

⨁

⨃

⨅

⨇

⨈

⨉

⨊

⫼

⫿


FAIL Test that non-stretchy largeops are bigger in display mode. assert_greater_than: width of largeop '⨇' expected a number greater than 6.640625 but got 2.640625
